EQUATIONS: Add modulo operator and round(), urand(), nrand() and srand() functions
The modulo operator x % y returnes the floating point remainder of
the division of x by y.

The round(x) functions returns the nearest integer value to a given
value.

The urand(min, max) and nrand(mean, sigma) return a random number
on a uniform and normal distribution respectively. The rands(s)
function can be used to set the seed for urand() and nrand().