SCRIPTS: Share the same variable array between all the equation parsers
The equation parser can now take a double array for the variable values
instead of allocating the array itself. This is used by the script parser
to share the same array between all the equation parser used. This
improves performances by decreasing the amount of allocation, and more
importantly by removing all the memcpy from the script parser array
to each equation parser arrays when evaluating the script. Some quick
profiling tests show an improvement of between 20% and 25% in the run
time.