SCRIPTS: Support single line if/else/while blocks
Before this change for all if/else/while blocks we needed
to use brackets. Now they are optional if the block only
contains one line.

So we can for example now use:
if (foo == 0)
  bar = 0;

But be careful as while the following is supported, it will
probably not do what you expect (the else is applied to the
first if and not the second one):
if (foo == 0)
  if (bar == 0)
    var = 0;
  else
    var = 1;

To try to mitigate this issue the following will result in
an error:
if (foo == 0)
  if (bar == 0) {
    var = 0;
  } else {
    var = 1;
  }