SCRIPTS: Detect use of 'else' not preceded by 'if' or 'if else'
This means that the following now gives an error:
if (foo == 1)
  bar = 1;
else
  bar = 5;
else
  bar = 7;

The old code was just passing it to the equation parser, which
then detected it as a script variable. So in some cases this did
not result in an error, but it did not do what was expected
either, and in other cases this resulted in a non-obvious error.