SCRIPTS: Fix two bugs when parsing if / if else blocks
The first bug occured when a 'else' statement followed a 'if else'
statement as the reformating done during the 'if else' processing
caused the 'else' to not be properly detected afterward.

The second bug was that 'if else' statement was not properly handled
if a line break occured between the 'if' and 'else'.

These bugs caused the following two scripts to result in an error
(unexpected 'else') despite being valid:

if (foo == 1)
  bar = 1;
else if (foo == 5)
  bar = 2;
else
  bar = -1;

if (foo == 1)
  bar = 1;
else
if (foo == 5)
  bar = 2;
else
  bar = -1;
due to the wrong reformating after parsing a